Move changes to individual models
