+++ model/line_2d.py
@@ -58,8 +58,8 @@ class SlvsLine2D(SlvsGenericEntity, PropertyGroup, Entity2D):
         self._batch = batch_for_shader(self._shader, "LINES", kwargs)
         self.is_dirty = False
 
-    def create_slvs_data(self, solvesys, group=Solver.group_fixed):
-        handle = solvesys.addLineSegment(self.p1.py_data, self.p2.py_data, group=group)
+    def create_slvs_data(self, solvesys):
+        handle = solvesys.add_line_2d(self.p1.py_data, self.p2.py_data, self.wp.py_data)
         self.py_data = handle
 
     def closest_picking_point(self, origin, view_vector):
